requirements.txt
streamlit
requests
pandas
scikit-learn
python-dotenv
plotly
joblib
